+++ yu-brm/src/main/java/com/yu/brm/service/impl/BrmMajorServiceImpl.java
@@ -1,11 +1,15 @@
 package com.yu.brm.service.impl;
 
 import java.util.List;
+import com.yu.common.exception.ServiceException;
 import com.yu.common.utils.DateUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import com.yu.brm.mapper.BrmMajorMapper;
+import com.yu.brm.mapper.BrmClassMapper;
 import com.yu.brm.domain.BrmMajor;
+import com.yu.brm.domain.BrmClass;
 import com.yu.brm.service.IBrmMajorService;
 
 /**
@@ -20,6 +24,9 @@ public class BrmMajorServiceImpl implements IBrmMajorService
     @Autowired
     private BrmMajorMapper brmMajorMapper;
 
+    @Autowired
+    private BrmClassMapper brmClassMapper;
+
     @Override
     public BrmMajor selectBrmMajorByMajorId(Long majorId)
     {
@@ -33,6 +40,7 @@ public class BrmMajorServiceImpl implements IBrmMajorService
     }
 
     @Override
+    @Transactional
     public int insertBrmMajor(BrmMajor brmMajor)
     {
         brmMajor.setCreateTime(DateUtils.getNowDate());
@@ -40,6 +48,7 @@ public class BrmMajorServiceImpl implements IBrmMajorService
     }
 
     @Override
+    @Transactional
     public int updateBrmMajor(BrmMajor brmMajor)
     {
         brmMajor.setUpdateTime(DateUtils.getNowDate());
@@ -47,14 +56,33 @@ public class BrmMajorServiceImpl implements IBrmMajorService
     }
 
     @Override
+    @Transactional
     public int deleteBrmMajorByMajorId(Long majorId)
     {
+        BrmClass query = new BrmClass();
+        query.setMajorId(majorId);
+        List<BrmClass> classes = brmClassMapper.selectBrmClassList(query);
+        if (classes != null && !classes.isEmpty())
+        {
+            throw new ServiceException("该专业下存在班级，不允许删除");
+        }
         return brmMajorMapper.deleteBrmMajorByMajorId(majorId);
     }
 
     @Override
+    @Transactional
     public int deleteBrmMajorByMajorIds(Long[] majorIds)
     {
+        for (Long majorId : majorIds)
+        {
+            BrmClass query = new BrmClass();
+            query.setMajorId(majorId);
+            List<BrmClass> classes = brmClassMapper.selectBrmClassList(query);
+            if (classes != null && !classes.isEmpty())
+            {
+                throw new ServiceException("该专业下存在班级，不允许删除");
+            }
+        }
         return brmMajorMapper.deleteBrmMajorByMajorIds(majorIds);
     }
 }
